Change Interface for login & Signup
Made changes to login page with regards to how it looks like.
I changed the background and rather placed the image side by side with the
login form. I did same for the signup too.

Further updates will be made in the future such as hover effects and focus effects